+++ frontend/src/App/component/DashboardPage.jsx
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import ProjectAvatar from './ProjectAvatar'
 import Axios from 'axios'
-
+import './Trello.css'
+import trelloLoad from './Trello.js'
 const useStyles = makeStyles((theme) => ({
     root: {
       display: 'flex',
@@ -35,6 +36,7 @@ function DashboardPage(props) {
         alert(e.response.status)
         console.error(e)
         })
+        trelloLoad();
     }, [])
 
     return (
@@ -43,9 +45,15 @@ function DashboardPage(props) {
             size = "small" 
             project={currentProject}
             />
-            <p>
+            <p></p>
             <h2>{currentProject ? currentProject.projectName : ""}</h2>
-            </p>
+            
+            <div className="base">
+              <div className="board">
+                <div className="board-lists" id="lists">
+                </div>
+              </div>
+            </div>
         </div>
     )
 }
