+++ frontend/src/App/component/Trello.css
@@ -5,10 +5,9 @@
     display: grid;
     grid-template-rows: max-content auto;
     grid-gap: 10px;
-    height: 500px;
-    width: 500px;
+    height: 800px;
+    width:50%;
     overflow: auto;
-    /*overflow-x: scroll;*/
     padding: 5px;
   }
   
@@ -83,7 +82,6 @@
     grid-auto-columns: 360px;
     grid-auto-flow: column;
     grid-gap: 8px;
-    overflow: scroll;
   }
   
   .board-list {
